cloud_functions/requirements.txt
requests
pandas
google-cloud-bigquery
pyarrow

